Fix rpc plugin option types
