+++ crates/atlaspack_plugin_rpc/src/nodejs/plugins/mod.rs
@@ -7,6 +7,7 @@ mod nodejs_rpc_reporter;
 mod nodejs_rpc_resolver;
 mod nodejs_rpc_runtime;
 mod nodejs_rpc_transformer;
+mod plugin_options;
 
 pub use self::nodejs_rpc_bundler::*;
 pub use self::nodejs_rpc_compressor::*;
